Add Follows Functionality
Added functionality for users to follow other users. Added method in
helpers to check of user already follows the current user.

Fix profile photo bug

Fixed the profile photo bug occuring on mobile version of the website.
Created a seperate partial for displaying photo if the user is not
current_user.

Fix Scroll On Sidebar

Fixed the scroll on sidebar by setting sidebars position to fixed and
putting sticky-top in navbar